requirements.txt
PySide6
pyserial>=3.0
pycryptodome